Streamlined tweaks; components now automatically support syntax
