+++ bridgetown-core/lib/bridgetown-core/configuration.rb
@@ -148,6 +148,7 @@ module Bridgetown
       dsl = ConfigurationDSL.new(scope: self, data: self)
       dsl.instance_variable_set(:@context, context)
       dsl.instance_exec(dsl, &init_init.block)
+      dsl._run_builtins!
       self.url = cached_url if cached_url # restore local development URL if need be
 
       setup_load_paths! appending: true
